+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * BREAKING: Asterisk translator now does NOT answer the call automatically when Output, Input or Record are used.
   * Change: Asterisk output now uses Playback rather than STREAM FILE
   * BREAKING: Asterisk output does not support interrupt digits
+  * Feature: The recordings dir is now checked for existence on startup, and logs an error if it is not there. Asterisk only.
 
   * Feature: Output component now exposes #recording and #recording_uri for easy access to record results.
   * Feature: Early media support for Asterisk, using Progress to start an early media session